1order-nproduct, ordermulti, 1order-1product

+++ ProductServiceApplication/src/main/java/com/store/dto/ProductDTO.java
@@ -3,12 +3,15 @@ package com.store.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductDTO {
     private String id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private int quantity;
 }
