+++ ProductServiceApplication/src/main/java/com/store/dto/ProductUpdateEvent.java
@@ -2,12 +2,15 @@ package com.store.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductUpdateEvent {
     private String productId;
     private String name;
-    private double price;
+    private BigDecimal price;
     private int quantity;
 }
